Show all workflow transitions in the history viewlet.
The standard workflow viewlet filters out automatic transitions. This is
undesirable for a full history overview since those transitions are a
proper part of the history. In particular we are looking at a use case
where an autmoatic transition is used to store the object creation
information (original author and creation date) which has to be visible.